refactor: centralize tape input projection
