+++ crates/agent-engine/src/runtime/turn_executor.rs
@@ -4,7 +4,7 @@ use std::time::Instant;
 use tokio_util::sync::CancellationToken;
 use tracing::{debug, error, info, warn};
 
-use crate::llm::{build_generation_request, project_tool_response_for_prompt};
+use crate::llm::build_generation_request;
 
 use super::agent_loop::{DeferredRuntimeAction, RuntimeLoopState};
 use super::compaction::{CompactionRequest, maybe_compact_context_with_cancel};
@@ -371,479 +371,6 @@ where
     }
 }
 
-fn responses_attachment_input_part(
-    hash: &str,
-    mime_type: &str,
-    metadata: &serde_json::Value,
-) -> serde_json::Value {
-    let image_like = mime_type.starts_with("image/");
-    if image_like {
-        if let Some(image_url) = metadata
-            .get("image_url")
-            .or_else(|| metadata.get("file_url"))
-            .or_else(|| metadata.get("url"))
-            .and_then(serde_json::Value::as_str)
-        {
-            return serde_json::json!({
-                "type": "input_image",
-                "image_url": image_url,
-            });
-        }
-        if let Some(file_id) = metadata.get("file_id").and_then(serde_json::Value::as_str) {
-            return serde_json::json!({
-                "type": "input_image",
-                "file_id": file_id,
-            });
-        }
-    }
-
-    if let Some(file_id) = metadata.get("file_id").and_then(serde_json::Value::as_str) {
-        return serde_json::json!({
-            "type": "input_file",
-            "file_id": file_id,
-        });
-    }
-
-    if let Some(file_url) = metadata
-        .get("file_url")
-        .or_else(|| metadata.get("url"))
-        .and_then(serde_json::Value::as_str)
-    {
-        return serde_json::json!({
-            "type": "input_file",
-            "file_url": file_url,
-        });
-    }
-
-    serde_json::json!({
-        "type": "input_text",
-        "text": format!("[attachment: {} ({})]", hash, mime_type),
-    })
-}
-
-fn chat_completions_attachment_content_part(
-    hash: &str,
-    mime_type: &str,
-    metadata: &serde_json::Value,
-) -> serde_json::Value {
-    if mime_type.starts_with("image/")
-        && let Some(image_url) = metadata
-            .get("image_url")
-            .or_else(|| metadata.get("file_url"))
-            .or_else(|| metadata.get("url"))
-            .and_then(serde_json::Value::as_str)
-    {
-        return serde_json::json!({
-            "type": "image_url",
-            "image_url": { "url": image_url },
-        });
-    }
-
-    if let Some(file_id) = metadata.get("file_id").and_then(serde_json::Value::as_str) {
-        return serde_json::json!({
-            "type": "file",
-            "file": { "file_id": file_id },
-        });
-    }
-
-    serde_json::json!({
-        "type": "text",
-        "text": format!("[attachment: {} ({})]", hash, mime_type),
-    })
-}
-
-fn anthropic_attachment_content_block(
-    hash: &str,
-    mime_type: &str,
-    metadata: &serde_json::Value,
-) -> serde_json::Value {
-    let block_type = if mime_type.starts_with("image/") {
-        "image"
-    } else {
-        "document"
-    };
-
-    if let Some(file_id) = metadata.get("file_id").and_then(serde_json::Value::as_str) {
-        let mut block = serde_json::json!({
-            "type": block_type,
-            "source": {
-                "type": "file",
-                "file_id": file_id,
-            },
-        });
-        if block_type == "document"
-            && let Some(title) = metadata.get("title").and_then(serde_json::Value::as_str)
-        {
-            block["title"] = serde_json::Value::String(title.to_string());
-        }
-        return block;
-    }
-
-    if let Some(url) = metadata
-        .get("file_url")
-        .or_else(|| metadata.get("image_url"))
-        .or_else(|| metadata.get("url"))
-        .and_then(serde_json::Value::as_str)
-    {
-        let mut block = serde_json::json!({
-            "type": block_type,
-            "source": {
-                "type": "url",
-                "url": url,
-            },
-        });
-        if block_type == "document"
-            && let Some(title) = metadata.get("title").and_then(serde_json::Value::as_str)
-        {
-            block["title"] = serde_json::Value::String(title.to_string());
-        }
-        return block;
-    }
-
-    serde_json::json!({
-        "type": "text",
-        "text": format!("[attachment: {} ({})]", hash, mime_type),
-    })
-}
-
-fn responses_message_content(parts: &[crate::tape::ContentPart]) -> Option<serde_json::Value> {
-    let needs_array = parts.iter().any(|part| {
-        !matches!(
-            part,
-            crate::tape::ContentPart::Text { .. } | crate::tape::ContentPart::Thinking { .. }
-        )
-    });
-
-    if !needs_array {
-        let text = crate::tape::parts_to_text(parts);
-        return (!text.trim().is_empty()).then_some(serde_json::Value::String(text));
-    }
-
-    let content_parts: Vec<serde_json::Value> = parts
-        .iter()
-        .filter_map(|part| match part {
-            crate::tape::ContentPart::Text { text } if !text.trim().is_empty() => {
-                Some(serde_json::json!({
-                    "type": "input_text",
-                    "text": text,
-                }))
-            }
-            crate::tape::ContentPart::Attachment {
-                hash,
-                mime_type,
-                metadata,
-            } => Some(responses_attachment_input_part(hash, mime_type, metadata)),
-            crate::tape::ContentPart::Structured { data } => Some(serde_json::json!({
-                "type": "input_text",
-                "text": data.to_string(),
-            })),
-            _ => None,
-        })
-        .collect();
-
-    (!content_parts.is_empty()).then_some(serde_json::Value::Array(content_parts))
-}
-
-fn chat_completions_message_content(
-    parts: &[crate::tape::ContentPart],
-) -> Option<serde_json::Value> {
-    let needs_array = parts.iter().any(|part| {
-        !matches!(
-            part,
-            crate::tape::ContentPart::Text { .. } | crate::tape::ContentPart::Thinking { .. }
-        )
-    });
-
-    if !needs_array {
-        let text = crate::tape::parts_to_text(parts);
-        return (!text.trim().is_empty()).then_some(serde_json::Value::String(text));
-    }
-
-    let content_parts: Vec<serde_json::Value> = parts
-        .iter()
-        .filter_map(|part| match part {
-            crate::tape::ContentPart::Text { text } if !text.trim().is_empty() => {
-                Some(serde_json::json!({
-                    "type": "text",
-                    "text": text,
-                }))
-            }
-            crate::tape::ContentPart::Attachment {
-                hash,
-                mime_type,
-                metadata,
-            } => Some(chat_completions_attachment_content_part(
-                hash, mime_type, metadata,
-            )),
-            crate::tape::ContentPart::Structured { data } => Some(serde_json::json!({
-                "type": "text",
-                "text": data.to_string(),
-            })),
-            _ => None,
-        })
-        .collect();
-
-    (!content_parts.is_empty()).then_some(serde_json::Value::Array(content_parts))
-}
-
-fn anthropic_message_content(parts: &[crate::tape::ContentPart]) -> Vec<serde_json::Value> {
-    parts
-        .iter()
-        .filter_map(|part| match part {
-            crate::tape::ContentPart::Text { text } if !text.trim().is_empty() => {
-                Some(serde_json::json!({
-                    "type": "text",
-                    "text": text,
-                }))
-            }
-            crate::tape::ContentPart::Thinking { text, signature } if !text.trim().is_empty() => {
-                let mut block = serde_json::json!({
-                    "type": "thinking",
-                    "thinking": text,
-                });
-                if let Some(signature) = signature
-                    .as_deref()
-                    .filter(|value| !value.trim().is_empty())
-                {
-                    block["signature"] = serde_json::Value::String(signature.to_string());
-                }
-                Some(block)
-            }
-            crate::tape::ContentPart::RedactedThinking { data } if !data.trim().is_empty() => {
-                Some(serde_json::json!({
-                    "type": "redacted_thinking",
-                    "data": data,
-                }))
-            }
-            crate::tape::ContentPart::Attachment {
-                hash,
-                mime_type,
-                metadata,
-            } => Some(anthropic_attachment_content_block(
-                hash, mime_type, metadata,
-            )),
-            crate::tape::ContentPart::Structured { data } => Some(serde_json::json!({
-                "type": "text",
-                "text": data.to_string(),
-            })),
-            _ => None,
-        })
-        .collect()
-}
-
-fn build_responses_input_items_from_tape(
-    messages: &[crate::agent_machine::Message],
-) -> Vec<serde_json::Value> {
-    let mut input = Vec::new();
-
-    for message in messages {
-        match message {
-            crate::agent_machine::Message::Tool { responses } => {
-                for response in responses {
-                    let projected_output = project_tool_response_for_prompt(&response.content);
-                    input.push(serde_json::json!({
-                        "type": "function_call_output",
-                        "call_id": response.id,
-                        "output": projected_output,
-                    }));
-                }
-            }
-            crate::agent_machine::Message::Assistant {
-                parts,
-                tool_requests,
-            } => {
-                if let Some(signature) = message.thinking_signature() {
-                    input.push(serde_json::json!({
-                        "type": "reasoning",
-                        "encrypted_content": signature,
-                    }));
-                }
-
-                if let Some(content) = responses_message_content(parts) {
-                    input.push(serde_json::json!({
-                        "role": "assistant",
-                        "content": content,
-                    }));
-                }
-
-                for tool_request in tool_requests {
-                    input.push(serde_json::json!({
-                        "type": "function_call",
-                        "call_id": tool_request.id,
-                        "name": tool_request.name,
-                        "arguments": tool_request.arguments.to_string(),
-                    }));
-                }
-            }
-            crate::agent_machine::Message::User { parts }
-            | crate::agent_machine::Message::System { parts }
-            | crate::agent_machine::Message::Context { parts } => {
-                if let Some(content) = responses_message_content(parts) {
-                    let role = match message.role() {
-                        crate::agent_machine::MessageRole::User => "user",
-                        _ => "developer",
-                    };
-                    input.push(serde_json::json!({
-                        "role": role,
-                        "content": content,
-                    }));
-                }
-            }
-        }
-    }
-
-    input
-}
-
-fn build_chat_completions_messages_from_tape(
-    messages: &[crate::agent_machine::Message],
-) -> Vec<serde_json::Value> {
-    let mut projected = Vec::new();
-
-    for message in messages {
-        match message {
-            crate::agent_machine::Message::Tool { responses } => {
-                for response in responses {
-                    let projected_content = project_tool_response_for_prompt(&response.content);
-                    projected.push(serde_json::json!({
-                        "role": "tool",
-                        "content": projected_content,
-                        "tool_call_id": response.id,
-                    }));
-                }
-            }
-            crate::agent_machine::Message::Assistant {
-                parts,
-                tool_requests,
-            } => {
-                let mut message_value = serde_json::json!({
-                    "role": "assistant",
-                });
-
-                if let Some(content) = chat_completions_message_content(parts) {
-                    message_value["content"] = content;
-                }
-                if let Some(thinking) = message.thinking_content() {
-                    message_value["reasoning_content"] = serde_json::Value::String(thinking);
-                }
-                if let Some(signature) = message.thinking_signature() {
-                    message_value["reasoning"] = serde_json::json!({
-                        "encrypted_content": signature,
-                    });
-                }
-                if !tool_requests.is_empty() {
-                    message_value["tool_calls"] = serde_json::Value::Array(
-                        tool_requests
-                            .iter()
-                            .map(|tool_request| {
-                                serde_json::json!({
-                                    "id": tool_request.id,
-                                    "type": "function",
-                                    "function": {
-                                        "name": tool_request.name,
-                                        "arguments": tool_request.arguments.to_string(),
-                                    },
-                                })
-                            })
-                            .collect(),
-                    );
-                }
-
-                projected.push(message_value);
-            }
-            crate::agent_machine::Message::User { parts } => {
-                if let Some(content) = chat_completions_message_content(parts) {
-                    projected.push(serde_json::json!({
-                        "role": "user",
-                        "content": content,
-                    }));
-                }
-            }
-            crate::agent_machine::Message::System { parts }
-            | crate::agent_machine::Message::Context { parts } => {
-                if let Some(content) = chat_completions_message_content(parts) {
-                    projected.push(serde_json::json!({
-                        "role": "developer",
-                        "content": content,
-                    }));
-                }
-            }
-        }
-    }
-
-    projected
-}
-
-fn build_anthropic_messages_from_tape(
-    messages: &[crate::agent_machine::Message],
-) -> Vec<serde_json::Value> {
-    let mut projected = Vec::new();
-    let mut known_tool_use_ids = std::collections::HashSet::new();
-
-    for message in messages {
-        match message {
-            crate::agent_machine::Message::Tool { responses } => {
-                for response in responses {
-                    let projected_content = project_tool_response_for_prompt(&response.content);
-                    let mut blocks = Vec::new();
-                    if known_tool_use_ids.contains(&response.id) {
-                        blocks.push(serde_json::json!({
-                            "type": "tool_result",
-                            "tool_use_id": response.id,
-                            "content": projected_content,
-                        }));
-                    } else if !projected_content.trim().is_empty() {
-                        blocks.push(serde_json::json!({
-                            "type": "text",
-                            "text": projected_content,
-                        }));
-                    }
-                    if !blocks.is_empty() {
-                        projected.push(serde_json::json!({
-                            "role": "user",
-                            "content": blocks,
-                        }));
-                    }
-                }
-            }
-            crate::agent_machine::Message::Assistant {
-                parts,
-                tool_requests,
-            } => {
-                let mut blocks = anthropic_message_content(parts);
-                for tool_request in tool_requests {
-                    known_tool_use_ids.insert(tool_request.id.clone());
-                    blocks.push(serde_json::json!({
-                        "type": "tool_use",
-                        "id": tool_request.id,
-                        "name": tool_request.name,
-                        "input": tool_request.arguments,
-                    }));
-                }
-                if !blocks.is_empty() {
-                    projected.push(serde_json::json!({
-                        "role": "assistant",
-                        "content": blocks,
-                    }));
-                }
-            }
-            crate::agent_machine::Message::User { parts } => {
-                let blocks = anthropic_message_content(parts);
-                if !blocks.is_empty() {
-                    projected.push(serde_json::json!({
-                        "role": "user",
-                        "content": blocks,
-                    }));
-                }
-            }
-            crate::agent_machine::Message::System { .. }
-            | crate::agent_machine::Message::Context { .. } => {}
-        }
-    }
-
-    projected
-}
-
 fn resolve_definition_persona_dirs(state: &RuntimeLoopState) -> Vec<std::path::PathBuf> {
     state.definition_persona_dirs.clone()
 }
@@ -1062,37 +589,26 @@ where
         let messages = prompt_view.messages;
         let raw_tape_messages = state.machine.tape.messages().to_vec();
         let mut previous_response_id: Option<String> = None;
-        let mut responses_input_items: Option<Vec<serde_json::Value>> = None;
-        let llm_messages = if responses_input_projection {
-            match supports_server_managed_continuation.then(|| {
-                resolve_responses_continuation(
-                    state,
-                    provider,
-                    context_revision,
-                    raw_tape_messages.len(),
-                )
-            }) {
-                Some(Some(continuation)) => {
-                    previous_response_id = Some(continuation.last_response_id);
-                    responses_input_items = Some(build_responses_input_items_from_tape(
-                        &raw_tape_messages[continuation.boundary_message_count..],
-                    ));
-                    state.project_generation_messages(
-                        &raw_tape_messages[continuation.boundary_message_count..],
+        let provider_input_messages: &[crate::agent_machine::Message] =
+            if responses_input_projection {
+                match supports_server_managed_continuation.then(|| {
+                    resolve_responses_continuation(
+                        state,
+                        provider,
+                        context_revision,
+                        raw_tape_messages.len(),
                     )
+                }) {
+                    Some(Some(continuation)) => {
+                        previous_response_id = Some(continuation.last_response_id);
+                        &raw_tape_messages[continuation.boundary_message_count..]
+                    }
+                    None | Some(None) => &messages,
                 }
-                None => {
-                    responses_input_items = Some(build_responses_input_items_from_tape(&messages));
-                    state.project_generation_messages(&messages)
-                }
-                Some(None) => {
-                    responses_input_items = Some(build_responses_input_items_from_tape(&messages));
-                    state.project_generation_messages(&messages)
-                }
-            }
-        } else {
-            state.project_generation_messages(&messages)
-        };
+            } else {
+                &messages
+            };
+        let llm_messages = crate::llm::project_messages(provider_input_messages, true);
         let llm_tools: Vec<crate::llm::ToolDefinition> = tools
             .iter()
             .map(|t| {
@@ -1114,29 +630,11 @@ where
         if let Some(instruction) = pending_guardrail_instruction.as_deref() {
             append_system_instruction(&mut request, instruction);
         }
-        if matches!(
+        request = crate::llm::with_provider_input(
+            request,
             provider_capabilities.instruction_role,
-            crate::llm::InstructionRole::Developer
-        ) {
-            request = request.with_extra_param(
-                "chat_completions_messages",
-                serde_json::Value::Array(build_chat_completions_messages_from_tape(&messages)),
-            );
-        } else if matches!(
-            provider_capabilities.instruction_role,
-            crate::llm::InstructionRole::AnthropicSystem
-        ) {
-            request = request.with_extra_param(
-                "anthropic_messages",
-                serde_json::Value::Array(build_anthropic_messages_from_tape(&messages)),
-            );
-        }
-        if let Some(responses_input_items) = responses_input_items {
-            request = request.with_extra_param(
-                "responses_input_items",
-                serde_json::Value::Array(responses_input_items),
-            );
-        }
+            provider_input_messages,
+        );
         if supports_provider_compaction
             && let Some(compact_threshold) = responses_server_managed_compact_threshold(state)
         {
@@ -1531,7 +1029,7 @@ mod tests {
         rollout::{RolloutItem, RolloutRecorder},
         runtime::{NamespaceRuntimeEnvironment, RuntimeConfig, TurnState},
         skills::{ResolvedCapabilityView, ScopedPackageDir, SkillScope},
-        tape::{ContentPart, Message, ToolRequest, ToolResponse},
+        tape::{ContentPart, Message, ToolRequest},
         tools::{Tool, ToolContext, ToolRegistry, ToolResult},
     };
     use alan_llm::{
@@ -3412,252 +2910,6 @@ description: {description}
             "reference context should stay in the neutral llmfs message list: {:?}",
             request.messages
         );
-    }
-
-    #[test]
-    fn test_build_responses_input_items_from_tape_projects_developer_role_and_attachments() {
-        let messages = vec![
-            crate::agent_machine::Message::Context {
-                parts: vec![ContentPart::text("Workspace context")],
-            },
-            crate::agent_machine::Message::User {
-                parts: vec![
-                    ContentPart::text("What is in this image?"),
-                    ContentPart::Attachment {
-                        hash: "img_hash".to_string(),
-                        mime_type: "image/png".to_string(),
-                        metadata: json!({
-                            "image_url": "https://example.com/cat.png"
-                        }),
-                    },
-                ],
-            },
-        ];
-
-        let items = build_responses_input_items_from_tape(&messages);
-        assert_eq!(
-            items,
-            vec![
-                json!({
-                    "role": "developer",
-                    "content": "Workspace context"
-                }),
-                json!({
-                    "role": "user",
-                    "content": [
-                        {
-                            "type": "input_text",
-                            "text": "What is in this image?"
-                        },
-                        {
-                            "type": "input_image",
-                            "image_url": "https://example.com/cat.png"
-                        }
-                    ]
-                })
-            ]
-        );
-    }
-
-    #[test]
-    fn test_build_chat_completions_messages_from_tape_projects_developer_role_and_attachments() {
-        let messages = vec![
-            crate::agent_machine::Message::Context {
-                parts: vec![ContentPart::text("Workspace context")],
-            },
-            crate::agent_machine::Message::User {
-                parts: vec![
-                    ContentPart::text("What is in this image?"),
-                    ContentPart::Attachment {
-                        hash: "img_hash".to_string(),
-                        mime_type: "image/png".to_string(),
-                        metadata: json!({
-                            "image_url": "https://example.com/cat.png"
-                        }),
-                    },
-                ],
-            },
-        ];
-
-        let projected = build_chat_completions_messages_from_tape(&messages);
-        assert_eq!(
-            projected,
-            vec![
-                json!({
-                    "role": "developer",
-                    "content": "Workspace context"
-                }),
-                json!({
-                    "role": "user",
-                    "content": [
-                        {
-                            "type": "text",
-                            "text": "What is in this image?"
-                        },
-                        {
-                            "type": "image_url",
-                            "image_url": {
-                                "url": "https://example.com/cat.png"
-                            }
-                        }
-                    ]
-                })
-            ]
-        );
-    }
-
-    #[test]
-    fn test_build_chat_completions_messages_from_tape_projects_file_url_image_attachments() {
-        let messages = vec![crate::agent_machine::Message::User {
-            parts: vec![
-                ContentPart::text("What is in this image?"),
-                ContentPart::Attachment {
-                    hash: "img_hash".to_string(),
-                    mime_type: "image/png".to_string(),
-                    metadata: json!({
-                        "file_url": "https://example.com/cat.png"
-                    }),
-                },
-            ],
-        }];
-
-        let projected = build_chat_completions_messages_from_tape(&messages);
-        assert_eq!(
-            projected,
-            vec![json!({
-                "role": "user",
-                "content": [
-                    {
-                        "type": "text",
-                        "text": "What is in this image?"
-                    },
-                    {
-                        "type": "image_url",
-                        "image_url": {
-                            "url": "https://example.com/cat.png"
-                        }
-                    }
-                ]
-            })]
-        );
-    }
-
-    #[test]
-    fn test_build_anthropic_messages_from_tape_projects_file_attachments() {
-        let messages = vec![crate::agent_machine::Message::User {
-            parts: vec![
-                ContentPart::text("Read this document"),
-                ContentPart::Attachment {
-                    hash: "doc_hash".to_string(),
-                    mime_type: "application/pdf".to_string(),
-                    metadata: json!({
-                        "file_id": "file_123",
-                        "title": "Spec"
-                    }),
-                },
-            ],
-        }];
-
-        let projected = build_anthropic_messages_from_tape(&messages);
-        assert_eq!(
-            projected,
-            vec![json!({
-                "role": "user",
-                "content": [
-                    {
-                        "type": "text",
-                        "text": "Read this document"
-                    },
-                    {
-                        "type": "document",
-                        "source": {
-                            "type": "file",
-                            "file_id": "file_123"
-                        },
-                        "title": "Spec"
-                    }
-                ]
-            })]
-        );
-    }
-
-    #[test]
-    fn test_build_responses_input_items_from_tape_caps_tool_payloads() {
-        let large_output = "x".repeat(40_000);
-        let messages = vec![crate::agent_machine::Message::Tool {
-            responses: vec![ToolResponse {
-                id: "call-1".to_string(),
-                content: vec![ContentPart::text(large_output.clone())],
-            }],
-        }];
-
-        let items = build_responses_input_items_from_tape(&messages);
-        let output = items[0]
-            .get("output")
-            .and_then(serde_json::Value::as_str)
-            .expect("responses item should contain string output");
-
-        assert_eq!(
-            output,
-            project_tool_response_for_prompt(&[ContentPart::text(large_output)])
-        );
-        assert!(output.len() <= 30_003);
-    }
-
-    #[test]
-    fn test_build_chat_completions_messages_from_tape_caps_tool_payloads() {
-        let large_output = "x".repeat(40_000);
-        let messages = vec![crate::agent_machine::Message::Tool {
-            responses: vec![ToolResponse {
-                id: "call-1".to_string(),
-                content: vec![ContentPart::text(large_output.clone())],
-            }],
-        }];
-
-        let projected = build_chat_completions_messages_from_tape(&messages);
-        let output = projected[0]
-            .get("content")
-            .and_then(serde_json::Value::as_str)
-            .expect("chat completions tool message should contain string content");
-
-        assert_eq!(
-            output,
-            project_tool_response_for_prompt(&[ContentPart::text(large_output)])
-        );
-        assert!(output.len() <= 30_003);
-    }
-
-    #[test]
-    fn test_build_anthropic_messages_from_tape_caps_tool_payloads() {
-        let large_output = "x".repeat(40_000);
-        let messages = vec![
-            crate::agent_machine::Message::Assistant {
-                parts: Vec::new(),
-                tool_requests: vec![ToolRequest {
-                    id: "call-1".to_string(),
-                    name: "tool".to_string(),
-                    arguments: json!({}),
-                }],
-            },
-            crate::agent_machine::Message::Tool {
-                responses: vec![ToolResponse {
-                    id: "call-1".to_string(),
-                    content: vec![ContentPart::text(large_output.clone())],
-                }],
-            },
-        ];
-
-        let projected = build_anthropic_messages_from_tape(&messages);
-        let output = projected[1]["content"][0]
-            .get("content")
-            .and_then(serde_json::Value::as_str)
-            .expect("anthropic tool_result should contain string content");
-
-        assert_eq!(
-            output,
-            project_tool_response_for_prompt(&[ContentPart::text(large_output)])
-        );
-        assert!(output.len() <= 30_003);
     }
 
     #[tokio::test]
