+++ crates/agent-engine/src/runtime/turn_support.rs
@@ -72,79 +72,6 @@ pub(super) fn normalize_tool_calls(
         .collect()
 }
 
-fn non_empty_trimmed(value: &str) -> Option<String> {
-    let trimmed = value.trim();
-    if trimmed.is_empty() {
-        None
-    } else {
-        Some(trimmed.to_string())
-    }
-}
-
-pub(super) fn project_messages_for_namespace(
-    messages: &[crate::agent_machine::Message],
-) -> Vec<crate::llm::Message> {
-    use crate::tape;
-
-    messages
-        .iter()
-        .flat_map(|message| match message {
-            tape::Message::Tool { responses } => responses
-                .iter()
-                .map(|response| crate::llm::Message {
-                    role: crate::llm::MessageRole::Tool,
-                    content: crate::llm::project_tool_response_for_prompt(&response.content),
-                    thinking: None,
-                    thinking_signature: None,
-                    redacted_thinking: None,
-                    tool_calls: None,
-                    tool_call_id: non_empty_trimmed(&response.id),
-                })
-                .collect::<Vec<_>>(),
-            _ => {
-                let role = match message.role() {
-                    tape::MessageRole::System => crate::llm::MessageRole::System,
-                    tape::MessageRole::Context => crate::llm::MessageRole::Context,
-                    tape::MessageRole::User => crate::llm::MessageRole::User,
-                    tape::MessageRole::Assistant => crate::llm::MessageRole::Assistant,
-                    tape::MessageRole::Tool => crate::llm::MessageRole::Tool,
-                };
-
-                vec![crate::llm::Message {
-                    role,
-                    content: message.non_thinking_text_content(),
-                    thinking: message.thinking_content(),
-                    thinking_signature: message.thinking_signature(),
-                    redacted_thinking: {
-                        let blocks = message.redacted_thinking_blocks();
-                        if blocks.is_empty() {
-                            None
-                        } else {
-                            Some(blocks)
-                        }
-                    },
-                    tool_calls: if message.tool_requests().is_empty() {
-                        None
-                    } else {
-                        Some(
-                            message
-                                .tool_requests()
-                                .iter()
-                                .map(|tool_request| crate::llm::ToolCall {
-                                    id: non_empty_trimmed(&tool_request.id),
-                                    name: tool_request.name.clone(),
-                                    arguments: tool_request.arguments.clone(),
-                                })
-                                .collect(),
-                        )
-                    },
-                    tool_call_id: None,
-                }]
-            }
-        })
-        .collect()
-}
-
 pub(super) fn tool_result_preview(value: &serde_json::Value) -> Option<String> {
     const MAX_PREVIEW_CHARS: usize = 160;
