+++ scripts/rust-source-size-baseline.txt
@@ -6,7 +6,7 @@ crates/agent-engine/src/agent_machine.rs 3661
 crates/agent-engine/src/config.rs 2556
 crates/agent-engine/src/policy.rs 1177
 crates/agent-engine/src/rollout.rs 2266
-crates/agent-engine/src/runtime/agent_loop.rs 2770
+crates/agent-engine/src/runtime/agent_loop.rs 2763
 crates/agent-engine/src/runtime/agent_loop/namespace_environment.rs 3747
 crates/agent-engine/src/runtime/child_agents.rs 4859
 crates/agent-engine/src/runtime/compaction.rs 1387
@@ -17,7 +17,7 @@ crates/agent-engine/src/runtime/prompt_cache.rs 2233
 crates/agent-engine/src/runtime/submission_handlers.rs 2981
 crates/agent-engine/src/runtime/tool_orchestrator.rs 3599
 crates/agent-engine/src/runtime/tool_policy.rs 1107
-crates/agent-engine/src/runtime/turn_executor.rs 6149
+crates/agent-engine/src/runtime/turn_executor.rs 5401
 crates/agent-engine/src/runtime/virtual_tools.rs 2755
 crates/agent-engine/src/runtime/virtual_tools_tests.rs 3680
 crates/agent-engine/src/skills/injector.rs 2490
